Rename testutil namespace
